tools/rag_and_search_tools/requirements.txt
google-cloud-discoveryengine
mcp
google-adk
uvicorn
starlette
python-dotenv